forceful crash of acceptance test jvm and store artifacts in circleci
