+++ acceptance-tests/tests/build.gradle
@@ -66,11 +66,17 @@ task acceptanceTest(type: Test) {
   description = 'Runs Besu acceptance tests.'
   group = 'verification'
 
+  jvmArgs "-XX:ErrorFile=${buildDir}/jvmErrorLogs/java_err_pid%p.log"
+
   testLogging {
     exceptionFormat = 'full'
     showStackTraces = true
     showStandardStreams = Boolean.getBoolean('acctests.showStandardStreams')
     showExceptions = true
     showCauses = true
+  }
+
+  doFirst {
+    mkdir "${buildDir}/jvmErrorLogs"
   }
 }
